Model/requirements.txt
absl-py==2.1.0
anyascii==0.3.2
astunparse==1.6.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
contractions==0.1.73
cycler==0.12.1
emoji==2.12.1
flatbuffers==24.3.25
fonttools==4.53.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
joblib==1.4.2
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
nltk==3.8.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyahocorasick==2.1.0
Pygments==2.18.0
pyparsing==3.1.2
pyspellchecker==0.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
textsearch==0.0.24
textattack
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
